\name{omixerRpm-package}
\alias{omixerRpm-package}
\alias{omixerRpm}
\docType{package}
\title{
\packageTitle{omixerRpm}
}
\description{
\packageDescription{omixerRpm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{omixerRpm}
\packageIndices{omixerRpm}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~

}
\author{
\packageAuthor{omixerRpm}

Maintainer: \packageMaintainer{omixerRpm}
}
\references{
~~ Literature or other references for background information ~~
omixerRpm was developed as part of GOmixer. If you use omixerRpm in your work please cite: 

Darzi, et al. Towards biome-specific analysis of meta-omics data, The ISME journal, 2015. http://www.nature.com/articles/ismej2015188

In case you use the bundled GMMs database, please also cite Vieira-Silva, et al. Species-function relationships shape ecological properties 
of the human gut microbiome, Nature Microbiology, 2016 .http://www.nature.com/articles/nmicrobiol201688

}
\keyword{ package , microbiome, gut, metagenomics, metabolic pathways }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
rpm()
}
